[Rename] : socketAtoms 파일명 변경
