+++ components/modules/ThankMixing.tsx
@@ -2,7 +2,7 @@
 
 import { useRecoilValue } from 'recoil';
 
-import { useCount } from '@/atoms/socketAtoms';
+import { useCount } from '@/atoms/useSocketAtoms';
 import MixingImg from 'public/images/mixing-img.svg';
 
 export default function ThankMixing() {
